Make sure Python 3.7 is supported for now
There are a lot of features that were introduced in Python 3.8
that are very useful and makes the code easier to read.

When Python 3.7 is EOL at the end of June 2023, we should increase
the minimum version to 3.8.

Added support for running all of the tests in Python 3.7 as well.