Win fixes. Save configaccess.log in debug mode
To make it easier to see which areas are accessed,
LL now saves a configaccess.log file on exit, if log
level is >=2. This shows a list of all config items that
have been read, written, or where an error occurred,
including access count.

Also fixed a couple of things to do with running in Windows.