Move more of the scheduling config to configdefs
This centralises the storage of scheduling information and
reduced redundancy in the code. It also makes it much easier
to add new scheduled tasks in the future.

This initial work is mainly to make everything work as it used
to, there is still a lot of things that could be optimized.