Fix no_umlauts issue using new general mechanism
The broader issue is that no_umlauts depends on the CONFIG['IMP_PREFLANG'] setting, and we want to avoid importing CONFIG in formatting.

To work around this, added a new ImportPrefs class that holds a copy of the preferences we need.

Also added a simple mechanism for automatically calling a class method when a config value changes - and use this to keep the new settings in formatter in sync.