Push remaining part of cleanup.py change
Now uses the location of the cleanup.py to find out where the basedir is, which seems more reliable.

Remaining problem: After deleting cherrypy, LL no longer starts cleanly. It gets to the right point, then fails because the import that was resolved earlier no longer works.

Restarting fixes it; it would be nice to fix.