Fix init script issues
Init.d script contained a space after the `--daemon` flag, which would ultimately become a double space when paired with any of the following flags. The double space would cause the `pgrep` in the `is_running()` function to not find any results, since the double space would be parsed out by the kernel but still exist in the `$DAEMON_OPTS` variable that is being used to grep. When `is_running()` fails the application cannot be stopped, since it doesn't believe it's running.

Additionally, on certain versions of Debian there is an issue when running `start-stop-daemon` with a `--pidfile` that is not owned by root. This causes the stop command to fail and an error to be displayed. Include the `--user $RUN_AS` flag as a workaround.